created creep-utils, started moving creep-related functions
